Added functionality to allocproc in proc.c to set parameters on process spawn. also added functionality to allocproc and wakeup1 in proc.c to avoid starvation. also created fill_pstat in proc.c. Scheduler is created but does not work.

+++ xv6/kernel/proc.h
@@ -76,8 +76,8 @@ struct proc {
   char name[16];               // Process name (debugging)
 
   int tickets;	// number of tickets assigned to this process
-  int pass;	// current pass value for the process
-  int stride;	// stride value
+  long pass;	// current pass value for the process
+  long stride;	// stride value
   int n_schedule;	// number of times chosen for scheduling
 };
